LPS-171715 search-experiences-web: Move fetch functions into a directory in utils
