+++ modules/dxp/apps/search-experiences/search-experiences-web/src/main/resources/META-INF/resources/sxp_blueprint_admin/js/utils/fetch/fetch_data.js
@@ -11,16 +11,15 @@
 
 import {fetch} from 'frontend-js-web';
 
-import {DEFAULT_ERROR} from './errorMessages';
+import {DEFAULT_ERROR} from '../errorMessages';
 
-const DEFAULT_HEADERS_OBJECT = {
+export const DEFAULT_HEADERS_OBJECT = {
 	'Accept': 'application/json',
 	'Accept-Language': Liferay.ThemeDisplay.getBCP47LanguageId(),
 	'Content-Type': 'application/json',
 };
 
 export const DEFAULT_HEADERS = new Headers(DEFAULT_HEADERS_OBJECT);
-export const DEFAULT_METHOD = 'GET';
 
 /**
  * A wrapper around frontend-js-web's `fetch` function with commonly used
@@ -31,10 +30,10 @@ export const DEFAULT_METHOD = 'GET';
  * @param {Object=} init An optional object containing custom configuration.
  * @return {Promise} A Promise that resolves to a Response object.
  */
-export function fetchData(resource = '', {headers, ...init} = {}) {
+export default function fetchData(resource = '', {headers, ...init} = {}) {
 	return fetch(resource, {
 		headers: new Headers({...DEFAULT_HEADERS_OBJECT, ...headers}),
-		method: DEFAULT_METHOD,
+		method: 'GET',
 		...init,
 	})
 		.then((response) => {
@@ -51,40 +50,4 @@ export function fetchData(resource = '', {headers, ...init} = {}) {
 
 			throw error;
 		});
-}
-
-/**
- * Modifies the url to include parameters.
- * @param {string} url The base url to fetch.
- * @param {Object} params The url parameters to be included in the request.
- * @returns {string} The modified url.
- */
-export function addParams(url, params) {
-	const fetchURL = new URL(url, Liferay.ThemeDisplay.getPortalURL());
-
-	Object.keys(params).forEach((key) => {
-		if (params[key] !== null) {
-			fetchURL.searchParams.append(key, params[key]);
-		}
-	});
-
-	return fetchURL;
-}
-
-/**
- * A wrapper function to fetch data for the preview sidebar. This was split into
- * a separate function primarily to make it easier to mock in tests.
- * @param {object} urlParameters The parameters to be added to the url.
- * @param {object} options Additional fetch options. For example, `{body: ...}`
- * @returns
- */
-export function fetchPreviewSearch(urlParameters, options) {
-	return fetch(
-		addParams('/o/search-experiences-rest/v1.0/search', urlParameters),
-		{
-			headers: DEFAULT_HEADERS,
-			method: 'POST',
-			...options,
-		}
-	);
 }
